% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordinateCleaner-package.R
\docType{data}
\name{aohi}
\alias{aohi}
\title{Artificial Hotspot Occurrence Inventory}
\source{
\url{https://doi.org/10.1111/jbi.14543}
}
\description{
A data frame with information on Artificial Hotspot Occurrence Inventory (AHOI)
 as available in Park et al 2022. For more details see reference.
}
\examples{

data("aohi")
}
\references{
Park, D. S., Xie, Y., Thammavong, H. T., Tulaiha, R., & Feng, X.
  (2023). Artificial Hotspot Occurrence Inventory (AHOI). Journal of
  Biogeography, 50, 441–449. https://doi.org/10.1111/jbi.14543
}
\keyword{gazetteers}
