% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fossils.R
\name{clean_fossils}
\alias{clean_fossils}
\alias{CleanCoordinatesFOS}
\title{Geographic and Temporal Cleaning of Records from Fossil Collections}
\usage{
clean_fossils(x, lon = "lng", lat = "lat", min_age = "min_ma",
  max_age = "max_ma", taxon = "accepted_name", tests = c("agesequal",
  "centroids", "equal", "gbif", "institutions", "spatiotemp", "temprange",
  "validity", "zeros"), countries = NULL, centroids_rad = 0.05,
  centroids_detail = "both", inst_rad = 0.001,
  outliers_method = "quantile", outliers_threshold = 5,
  outliers_size = 7, outliers_replicates = 5, zeros_rad = 0.5,
  centroids_ref = NULL, country_ref = NULL, inst_ref = NULL,
  value = "spatialvalid", verbose = TRUE, report = FALSE)
}
\arguments{
\item{x}{a data.frame. Containing geographical coordinates and species
names.}

\item{lon}{a character string. The column with the longitude coordinates.
Default = \dQuote{decimallongitude}.}

\item{lat}{a character string. The column with the longitude coordinates.
Default = \dQuote{decimallatitude}.}

\item{min_age}{a character string. The column with the minimum age. Default
= \dQuote{min_ma}.}

\item{max_age}{a character string. The column with the maximum age. Default
= \dQuote{max_ma}.}

\item{taxon}{a character string. The column with the taxon name. If
\dQuote{}, searches for outliers over the entire dataset, otherwise per
specified taxon. Default = \dQuote{accepted_name}.}

\item{tests}{a vector of character strings, indicating which tests to run.
See details for all tests available. Default = c("centroids",
"equal", "gbif", "institutions", "temprange", "spatiotemp", "agesequal", "zeros")}

\item{countries}{a character string. A vector of the same length as rows in
x, with country information for each record in ISO3 format.  If missing, the
countries test is skipped.}

\item{centroids_rad}{numeric. The side length of the rectangle around
country centroids in degrees. Default = 0.01.}

\item{centroids_detail}{a \code{character string}. If set to
\sQuote{country} only country (adm-0) centroids are tested, if set to
\sQuote{provinces} only province (adm-1) centroids are tested.  Default =
\sQuote{both}.}

\item{inst_rad}{numeric. The radius around biodiversity institutions
coordinates in degrees. Default = 0.001.}

\item{outliers_method}{The method used for outlier testing. See details.}

\item{outliers_threshold}{numerical.  The multiplier for the interquantile
range for outlier detection. The higher the number, the more conservative
the outlier tests.  See \code{\link{cf_outl}} for details. Default = 3.}

\item{outliers_size}{numerical.  The minimum number of records in a dataset
to run the taxon-specific outlier test.  Default = 7.}

\item{outliers_replicates}{numeric. The number of replications for the
distance matrix calculation. See details.  Default = 5.}

\item{zeros_rad}{numeric. The radius around 0/0 in degrees. Default = 0.5.}

\item{centroids_ref}{a \code{data.frame} with alternative reference data for
the centroid test. If missing, the \code{countryref} dataset is used.
Alternatives must be identical in structure.}

\item{country_ref}{a \code{SpatialPolygonsDataFrame} as alternative
reference for the countries test. If missing, the
\code{rnaturalearth:ne_countries('medium')} dataset is used.}

\item{inst_ref}{a \code{data.frame} with alternative reference data for the
biodiversity institution test. If missing, the \code{institutions} dataset
is used.  Alternatives must be identical in structure.}

\item{value}{a character string defining the output value. See the value
section for details. one of \sQuote{spatialvalid}, \sQuote{summary},
\sQuote{cleaned}. Default = \sQuote{\code{spatialvalid}}.}

\item{verbose}{logical. If TRUE reports the name of the test and the number
of records flagged}

\item{report}{logical or character.  If TRUE a report file is written to the
working directory, summarizing the cleaning results. If a character, the
path to which the file should be written.  Default = FALSE.}
}
\value{
Depending on the output argument: \describe{
\item{list("spatialvalid")}{an object of class \code{spatialvalid} with one
column for each test. TRUE = clean coordinate, FALSE = potentially
problematic coordinates.  The summary column is FALSE if any test flagged
the respective coordinate.} \item{list("flagged")}{a logical vector with the
same order as the input data summarizing the results of all test. TRUE =
clean coordinate, FALSE = potentially problematic (= at least one test
failed).} \item{list("cleaned")}{a \code{data.frame} of cleaned coordinates
if \code{species = NULL} or a \code{data.frame} with cleaned coordinates and
species ID otherwise} }
}
\description{
Cleaning records by multiple empirical tests to flag potentially erroneous
coordinates and time-spans, addressing issues common in fossil collection
databases. Individual tests can be activated via the tests argument:
}
\details{
\itemize{
\item agesequal tests for equal minimum and maximum age.
\item centroids tests a radius around country centroids.
The radius is \code{centroids_rad}.
\item countries tests if coordinates are from the
country indicated in the country column.  \emph{Switched off by default.}
\item equal tests for equal absolute longitude and latitude.
\item gbif tests a one-degree radius around the GBIF
headquarters in Copenhagen, Denmark.
\item institutions tests a radius around known
biodiversity institutions from \code{instiutions}. The radius is
\code{inst_rad}.
\item spatiotemp test for records which are outlier in time and space. See below for details.
\item temprange tests for records with unexpectedly large temporal ranges,
using a quantile-based outlier test.
\item validity checks if coordinates correspond to a lat/lon coordinate reference system.
This test is always on, since all records need to pass for any other test to run.
\item zeros tests for plain zeros, equal latitude and
longitude and a radius around the point 0/0. The radius is \code{zeros_rad}.
The outlier detection in \sQuote{spatiotemp} is based on an interquantile range test. In a first
step a distance matrix of geographic distances among all records is
calculate. Subsequently a similar distance matrix of temporal distances
among all records is calculated based on a single point selected by random
between the minimum and maximum age for each record. The mean distance for
each point to all neighbours is calculated for both matrices and spatial and
temporal distances are scaled to the same range. The sum of these distanced
is then tested against the interquantile range and flagged as an outlier if
$x > IQR(x) + q_75 * mltpl$. The test is replicated \sQuote{replicates}
times, to account for temporal uncertainty. Records are flagged as outliers
if they are flagged by a fraction of more than \sQuote{flag_thresh}
replicates. Only datasets/taxa comprising more than \sQuote{size.thresh}
records are tested. Note that geographic distances are calculated as
geospheric distances for datasets (or taxa) with less than 10,000 records
and approximated as Euclidean distances for datasets/taxa with 10,000 to
25,000 records. Datasets/taxa comprising more than 25,000 records are
skipped.
}
}
\note{
Always tests for coordinate validity: non-numeric or missing
coordinates and coordinates exceeding the global extent (lon/lat, WGS84).

See \url{https://azizka.github.io/CoordinateCleaner/} for more details
and tutorials.
}
\examples{

minages <- runif(250, 0, 65)
exmpl <- data.frame(accepted_name = sample(letters, size = 250, replace = TRUE),
                    lng = runif(250, min = 42, max = 51),
                    lat = runif(250, min = -26, max = -11),
                    min_ma = minages,
                    max_ma = minages + runif(250, 0.1, 65))

test <- clean_fossils(x = exmpl)

summary(test)

}
\seealso{
Other Wrapper functions: \code{\link{clean_coordinates}},
  \code{\link{clean_dataset}}
}
\concept{Wrapper functions}
\keyword{Coordinate}
\keyword{Fossil}
\keyword{Temporal}
\keyword{cleaning}
