% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordinateCleaner-package.R
\name{CoordinateCleaner-defunct}
\alias{CoordinateCleaner-defunct}
\title{Defunct functions in CoordinateCleaener}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
 \item \code{\link{CleanCoordinates}}: This function is deprecated, and will
 be removed in the next version of this package. Use \code{\link{clean_coordinates}}
 instead
 \item \code{\link{CleanCoordinatesDS}}: This function is deprecated, and will
 be removed in the next version of this package. Use \code{\link{clean_dataset}}
 instead
 \item \code{\link{CleanCoordinatesFOS}}: This function is deprecated, and will
 be removed in the next version of this package. Use \code{\link{clean_fossils}}
 instead
}
}
