\name{CoordinateCleaner-package}
\alias{CoordinateCleaner-package}
\alias{CoordinateCleaner}
\docType{package}
\title{
\packageTitle{CoordinateCleaner}
}
\description{
\packageDescription{CoordinateCleaner}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CoordinateCleaner}
\packageIndices{CoordinateCleaner}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CoordinateCleaner}

Maintainer: \packageMaintainer{CoordinateCleaner}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
