% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordinateCleaner-package.R
\docType{data}
\name{countryref}
\alias{countryref}
\title{Country Centroids and Country Capitals}
\format{A data frame with 249 observations on 7 variables.}
\source{
CENTRAL INTELLIGENCE AGENCY (2014) \emph{The World Factbook},
Washington, DC.

\url{http://opengeocode.org/download/cow.php}
}
\description{
A \code{data.frame} with coordinates of country centroids and country
capitals as reference for the \code{\link{clean_coordinates}} function.
Coordinates are based on the Central Intelligence Agency World Factbook as
provided at \url{http://opengeocode.org/download/cow.php}.
}
\examples{

data(countryref)

}
\keyword{gazetteers}
