% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordinateCleaner-package.R
\docType{data}
\name{institutions}
\alias{institutions}
\title{Global Locations of Biodiversity Institutions}
\format{
A data frame with 12170 observations on 12 variables.
}
\source{
Compiled from various sources: \itemize{ \item Global Biodiversity
Information Facility \url{https://www.gbif.org/} \item Wikipedia
\url{https://www.wikipedia.org/} \item Geonames \url{https://www.geonames.org/} \item The Global
Registry of Biodiversity Repositories \item Index
Herbariorum \url{https://sweetgum.nybg.org/science/ih/}
\item Botanic Gardens Conservation International \url{https://www.bgci.org/}
}
}
\description{
A global gazetteer for biodiversity institutions from various sources,
including zoos, museums, botanical gardens, GBIF contributors, herbaria,
university collections.
}
\examples{

data(institutions)
str(institutions)

}
\keyword{gazetteers}
