% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.spatialvalid.R
\name{is.spatialvalid}
\alias{is.spatialvalid}
\title{Check spatialvalid object}
\usage{
is.spatialvalid(x)
}
\arguments{
\item{x}{the object to be tested}
}
\value{
returns \code{TRUE} if its argument is a spatialvalid
}
\description{
Test if its argument is a spatialvalid object
}
\keyword{Check}
