% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordinateCleaner-package.R
\docType{data}
\name{pbdb_example}
\alias{pbdb_example}
\title{Example data from the Paleobiologydatabase}
\format{
A data frame with 5000 observations on 36 variables.
}
\source{
\itemize{ 
\item The Paleobiology database \url{https://paleobiodb.org/} 
\item Sara Varela, Javier Gonzalez Hernandez and Luciano Fabris Sgarbi (2016). 
paleobioDB: Download and Process Data from the Paleobiology Database. 
R package version 0.5.0. \url{https://CRAN.R-project.org/package=paleobioDB}.
}
}
\description{
A dataset of 5000 flowering plant fossil occurrences as example for data of the paleobiology Database, downloaded using the paleobioDB packages as specified in the vignette \dQuote{Cleaning_PBDB_fossils_with_CoordinateCleaner}.
}
\examples{

data(institutions)
str(institutions)

}
\keyword{gazetteers}
