% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordinateCleaner-package.R
\docType{data}
\name{capitals}
\alias{capitals}
\title{Global Capital Locations}
\format{A data frame with 225 observations on the following 4 variables.
\describe{\item{ISO3}{a factor, ISO-3 country code.}
\item{capital}{a factor, capital names.} \item{longitude}{a
numeric vector.} \item{latitude}{a numeric vector.} }}
\source{
CENTRAL INTELLIGENCE AGENCY (2014) \emph{The World Factbook},
Washington, DC.

\url{https://www.cia.gov/library/publications/the-world-factbook/}
}
\description{
A gazetteer of global capital coordinates.
}
\examples{

data(capitals)
str(capitals)

}
\keyword{gazetteers}
