% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_coun.R
\name{cc_coun}
\alias{cc_coun}
\title{Identify Coordinates Outside their Reported Country}
\usage{
cc_coun(
  x,
  lon = "decimallongitude",
  lat = "decimallatitude",
  iso3 = "countrycode",
  value = "clean",
  ref = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{data.frame. Containing geographical coordinates and species
names.}

\item{lon}{character string. The column with the longitude coordinates.
Default = \dQuote{decimallongitude}.}

\item{lat}{character string. The column with the latitude coordinates.
Default = \dQuote{decimallatitude}.}

\item{iso3}{a character string. The column with the country assignment of
each record in three letter ISO code. Default = \dQuote{countrycode}.}

\item{value}{character string.  Defining the output value. See value.}

\item{ref}{a SpatialPolygonsDataFrame. Providing the geographic gazetteer.
Can be any SpatialPolygonsDataFrame, but the structure must be identical to
\code{rnaturalearth::ne_countries(scale = "medium")}.  
Default = \code{rnaturalearth::ne_countries(scale = "medium")}}

\item{verbose}{logical. If TRUE reports the name of the test and the number
of records flagged.}
}
\value{
Depending on the \sQuote{value} argument, either a \code{data.frame}
containing the records considered correct by the test (\dQuote{clean}) or a
logical vector (\dQuote{flagged}), with TRUE = test passed and FALSE = test failed/potentially
problematic . Default = \dQuote{clean}.
}
\description{
Removes or flags mismatches between geographic coordinates and additional country
information (usually this information is reliably reported with specimens).
Such a mismatch can occur for example, if latitude and longitude are
switched.
}
\note{
With the default reference, records are flagged if they fall 
outside the terrestrial territory of countries, hence records in territorial waters might be flagged. 
See \url{https://ropensci.github.io/CoordinateCleaner/} for more
details and tutorials.
}
\examples{

\dontrun{
x <- data.frame(species = letters[1:10], 
                decimallongitude = runif(100, -20, 30), 
                decimallatitude = runif(100, 35,60),
                countrycode = "RUS")

cc_coun(x, value = "flagged")#non-terrestrial records are flagged as wrong. 
}

}
\seealso{
Other Coordinates: 
\code{\link{cc_cap}()},
\code{\link{cc_cen}()},
\code{\link{cc_dupl}()},
\code{\link{cc_equ}()},
\code{\link{cc_gbif}()},
\code{\link{cc_inst}()},
\code{\link{cc_iucn}()},
\code{\link{cc_outl}()},
\code{\link{cc_sea}()},
\code{\link{cc_urb}()},
\code{\link{cc_val}()},
\code{\link{cc_zero}()}
}
\concept{Coordinates}
\keyword{Coordinate}
\keyword{cleaning}
