% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordinateCleaner-package.R
\docType{package}
\name{CoordinateCleaner-package}
\alias{CoordinateCleaner-package}
\alias{CoordinateCleaner}
\title{c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_title(\"#1\")}",
"CoordinateCleaner")\Sexpr{tools:::Rd_package_title("CoordinateCleaner")}}
\description{
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_description(\"#1\")}",
"CoordinateCleaner")\Sexpr{tools:::Rd_package_description("CoordinateCleaner")}
}
\details{
The DESCRIPTION file:
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_DESCRIPTION(\"#1\")}",
"CoordinateCleaner")\Sexpr{tools:::Rd_package_DESCRIPTION("CoordinateCleaner")}
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_indices(\"#1\")}",
"CoordinateCleaner")\Sexpr{tools:::Rd_package_indices("CoordinateCleaner")}
}
\author{
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_author(\"#1\")}",
"CoordinateCleaner")\Sexpr{tools:::Rd_package_author("CoordinateCleaner")}

Maintainer:
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_maintainer(\"#1\")}",
"CoordinateCleaner")\Sexpr{tools:::Rd_package_maintainer("CoordinateCleaner")}
}
