\name{CoordinateCleaner-package}
\alias{CoordinateCleaner-package}
\alias{CoordinateCleaner}
\docType{package}
\title{
\packageTitle{CoordinateCleaner}
}
\description{
\packageDescription{CoordinateCleaner}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{CoordinateCleaner}
\packageIndices{CoordinateCleaner}
}
\author{
\packageAuthor{CoordinateCleaner}

Maintainer: \packageMaintainer{CoordinateCleaner}
}
