% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordinateCleaner-package.R
\docType{data}
\name{landmass}
\alias{landmass}
\title{Global Coastlines}
\source{
\url{http://www.naturalearthdata.com/downloads/10m-physical-vectors/}
}
\description{
A \code{SpatialPolygonsDataFrame} with global coastlines.
}
\note{
Most of the times it might be desirable to only flag records far away
from the coast as problematic rather than those close to the coastline
(which might be due to disagreements in coastlines, or low GPS uncertainty).
For these cases, there is a alternative coastline reference buffered by one
degree available at
\url{https://github.com/azizka/CoordinateCleaner/tree/master/extra_gazetteers}.
}
\examples{

data("landmass")

}
\keyword{gazetteers}
